backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
python-dotenv
psycopg2-binary
